// frontend/src/api/model/character.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterFranchiseId } from './characterFranchiseId';
import type { CharacterFranchise } from './characterFranchise';

export interface Character {
  /** The name of the character. */
  name: string;
  /** The ID of the franchise this character belongs to. */
  franchise_id?: CharacterFranchiseId;
  /** Unique database identifier for the character. */
  id: number;
  /** The franchise this character belongs to. */
  franchise?: CharacterFranchise;
  /** Number of sets featuring this character. */
  set_count?: number;
}